+++ autoscaling.tf
@@ -1,15 +1,15 @@
 resource "aws_appautoscaling_target" "table_read" {
-  count = var.create_table && length(var.autoscaling_read) > 0 ? 1 : 0
+  count = var.create_table && var.autoscaling_enabled && length(var.autoscaling_read) > 0 ? 1 : 0
 
   max_capacity       = var.autoscaling_read["max_capacity"]
   min_capacity       = var.read_capacity
-  resource_id        = "table/${aws_dynamodb_table.this[0].name}"
+  resource_id        = "table/${aws_dynamodb_table.autoscaled[0].name}"
   scalable_dimension = "dynamodb:table:ReadCapacityUnits"
   service_namespace  = "dynamodb"
 }
 
 resource "aws_appautoscaling_policy" "table_read_policy" {
-  count = var.create_table && length(var.autoscaling_read) > 0 ? 1 : 0
+  count = var.create_table && var.autoscaling_enabled && length(var.autoscaling_read) > 0 ? 1 : 0
 
   name               = "DynamoDBReadCapacityUtilization:${aws_appautoscaling_target.table_read[0].resource_id}"
   policy_type        = "TargetTrackingScaling"
@@ -29,17 +29,17 @@ resource "aws_appautoscaling_policy" "table_read_policy" {
 }
 
 resource "aws_appautoscaling_target" "table_write" {
-  count = var.create_table && length(var.autoscaling_write) > 0 ? 1 : 0
+  count = var.create_table && var.autoscaling_enabled && length(var.autoscaling_write) > 0 ? 1 : 0
 
   max_capacity       = var.autoscaling_write["max_capacity"]
   min_capacity       = var.write_capacity
-  resource_id        = "table/${aws_dynamodb_table.this[0].name}"
+  resource_id        = "table/${aws_dynamodb_table.autoscaled[0].name}"
   scalable_dimension = "dynamodb:table:WriteCapacityUnits"
   service_namespace  = "dynamodb"
 }
 
 resource "aws_appautoscaling_policy" "table_write_policy" {
-  count = var.create_table && length(var.autoscaling_write) > 0 ? 1 : 0
+  count = var.create_table && var.autoscaling_enabled && length(var.autoscaling_write) > 0 ? 1 : 0
 
   name               = "DynamoDBWriteCapacityUtilization:${aws_appautoscaling_target.table_write[0].resource_id}"
   policy_type        = "TargetTrackingScaling"
@@ -59,17 +59,17 @@ resource "aws_appautoscaling_policy" "table_write_policy" {
 }
 
 resource "aws_appautoscaling_target" "index_read" {
-  for_each = var.create_table ? var.autoscaling_indexes : {}
+  for_each = var.create_table && var.autoscaling_enabled ? var.autoscaling_indexes : {}
 
   max_capacity       = each.value["read_max_capacity"]
   min_capacity       = each.value["read_min_capacity"]
-  resource_id        = "table/${aws_dynamodb_table.this[0].name}/index/${each.key}"
+  resource_id        = "table/${aws_dynamodb_table.autoscaled[0].name}/index/${each.key}"
   scalable_dimension = "dynamodb:index:ReadCapacityUnits"
   service_namespace  = "dynamodb"
 }
 
 resource "aws_appautoscaling_policy" "index_read_policy" {
-  for_each = var.create_table ? var.autoscaling_indexes : {}
+  for_each = var.create_table && var.autoscaling_enabled ? var.autoscaling_indexes : {}
 
   name               = "DynamoDBReadCapacityUtilization:${aws_appautoscaling_target.index_read[each.key].resource_id}"
   policy_type        = "TargetTrackingScaling"
@@ -89,17 +89,17 @@ resource "aws_appautoscaling_policy" "index_read_policy" {
 }
 
 resource "aws_appautoscaling_target" "index_write" {
-  for_each = var.create_table ? var.autoscaling_indexes : {}
+  for_each = var.create_table && var.autoscaling_enabled ? var.autoscaling_indexes : {}
 
   max_capacity       = each.value["write_max_capacity"]
   min_capacity       = each.value["write_min_capacity"]
-  resource_id        = "table/${aws_dynamodb_table.this[0].name}/index/${each.key}"
+  resource_id        = "table/${aws_dynamodb_table.autoscaled[0].name}/index/${each.key}"
   scalable_dimension = "dynamodb:index:WriteCapacityUnits"
   service_namespace  = "dynamodb"
 }
 
 resource "aws_appautoscaling_policy" "index_write_policy" {
-  for_each = var.create_table ? var.autoscaling_indexes : {}
+  for_each = var.create_table && var.autoscaling_enabled ? var.autoscaling_indexes : {}
 
   name               = "DynamoDBWriteCapacityUtilization:${aws_appautoscaling_target.index_write[each.key].resource_id}"
   policy_type        = "TargetTrackingScaling"