+++ main.tf
@@ -1,5 +1,5 @@
 resource "aws_dynamodb_table" "this" {
-  count = var.create_table ? 1 : 0
+  count = var.create_table && !var.autoscaling_enabled ? 1 : 0
 
   name             = var.name
   billing_mode     = var.billing_mode
@@ -87,5 +87,92 @@ resource "aws_dynamodb_table" "this" {
       read_capacity,
       write_capacity
     ]
+  }
+}
+
+resource "aws_dynamodb_table" "autoscaled" {
+  count = var.create_table && var.autoscaling_enabled ? 1 : 0
+
+  name             = var.name
+  billing_mode     = var.billing_mode
+  hash_key         = var.hash_key
+  range_key        = var.range_key
+  read_capacity    = var.read_capacity
+  write_capacity   = var.write_capacity
+  stream_enabled   = var.stream_enabled
+  stream_view_type = var.stream_view_type
+
+  ttl {
+    enabled        = var.ttl_enabled
+    attribute_name = var.ttl_attribute_name
+  }
+
+  point_in_time_recovery {
+    enabled = var.point_in_time_recovery_enabled
+  }
+
+  dynamic "attribute" {
+    for_each = var.attributes
+
+    content {
+      name = attribute.value.name
+      type = attribute.value.type
+    }
+  }
+
+  dynamic "local_secondary_index" {
+    for_each = var.local_secondary_indexes
+
+    content {
+      name               = local_secondary_index.value.name
+      range_key          = local_secondary_index.value.range_key
+      projection_type    = local_secondary_index.value.projection_type
+      non_key_attributes = lookup(local_secondary_index.value, "non_key_attributes", null)
+    }
+  }
+
+  dynamic "global_secondary_index" {
+    for_each = var.global_secondary_indexes
+
+    content {
+      name               = global_secondary_index.value.name
+      hash_key           = global_secondary_index.value.hash_key
+      projection_type    = global_secondary_index.value.projection_type
+      range_key          = lookup(global_secondary_index.value, "range_key", null)
+      read_capacity      = lookup(global_secondary_index.value, "read_capacity", null)
+      write_capacity     = lookup(global_secondary_index.value, "write_capacity", null)
+      non_key_attributes = lookup(global_secondary_index.value, "non_key_attributes", null)
+    }
+  }
+
+  dynamic "replica" {
+    for_each = var.replica_regions
+
+    content {
+      region_name = replica.value.region_name
+      kms_key_arn = lookup(replica.value, "kms_key_arn", null)
+    }
+  }
+
+  server_side_encryption {
+    enabled     = var.server_side_encryption_enabled
+    kms_key_arn = var.server_side_encryption_kms_key_arn
+  }
+
+  tags = merge(
+    var.tags,
+    {
+      "Name" = format("%s", var.name)
+    },
+  )
+
+  timeouts {
+    create = lookup(var.timeouts, "create", null)
+    delete = lookup(var.timeouts, "delete", null)
+    update = lookup(var.timeouts, "update", null)
+  }
+
+  lifecycle {
+    ignore_changes = [read_capacity, write_capacity]
   }
 }